+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/model/CommerceProductDefinitionOptionValueRelSoap.java
@@ -174,11 +174,11 @@ public class CommerceProductDefinitionOptionValueRelSoap implements Serializable
 		_title = title;
 	}
 
-	public long getPriority() {
+	public int getPriority() {
 		return _priority;
 	}
 
-	public void setPriority(long priority) {
+	public void setPriority(int priority) {
 		_priority = priority;
 	}
 
@@ -191,5 +191,5 @@ public class CommerceProductDefinitionOptionValueRelSoap implements Serializable
 	private Date _modifiedDate;
 	private long _commerceProductDefinitionOptionRelId;
 	private String _title;
-	private long _priority;
+	private int _priority;
 }
